[Feat] API 서비스 모듈 통합
