--- src/components/modals/containers/FileUploadService.jsx
@@ -1,17 +1,0 @@
-import axios from 'axios';
-
-const FileUploadService = {
-  uploadFile: (detailCategoryId, file) => {
-    const formData = new FormData();
-    formData.append('file', file);
-    formData.append('detailCategoryId', detailCategoryId);
-
-    return axios.post('http://54.166.160.145:8080/article-file', formData, {
-      headers: {
-        'Content-Type': 'multipart/form-data'
-      }
-    });
-  }
-};
-
-export default FileUploadService;
